Rename bcm.* to bus.*, compile .c with dune
Renamed the bcm.c/h/ml files to bus.* and made minor modifications.
The .c files are now compiled with dune instead of relying on a separate
Makefile.

+++ lib/bus.c
@@ -1,5 +1,5 @@
 #include <bcm2835.h>
-#include "bcm.h"
+#include "bus.h"
 
 #define CS 8
 #define HRDY 24
@@ -21,7 +21,7 @@ void closeBus()
   bcm2835_gpio_write(CS, HIGH);
 }
 
-int transfer(int value)
+uint16_t transfer(uint16_t value)
 {
   value = value & 0xFFFF;
   int retVal = 0x0000;
